test(hll): focus lower-bound regression coverage
